proj: initial commit
This commit initializes the git repository for the amclient.

At this moment the client connects to an amserver, and upon pressing Space, it
sends a StartGame message to the server, which is answered back, and printed to
the console.

This is the start of the client-server architecture for Ars Militaris. From now,
development will be made in these two projects (amclient and amserver).