feat: wait turn synchronization
This commit marks the achievement of basic client-server synchronization for the
wait turn mechanic, and also the concept of having the player get its turn from
the server.

Some inconsistent behavior was found when implementing the functionality, in
that sometimes the server or client messages won't be received in the required
time. This has resulted most noticeably in bug #3.

Still, development will continue on.