feat: change wait turn behavior
This commit changes the bevahor of the wait_turn_system and its synchronization
with the server.

Now, the wait_turn_system only runs on the server, and, when, complete, it will
send the updated WTs to the client.

The client now knows, when receiving the PlayerTurn message, if it is its turn
or not, by querying the ClientData resource.

If it is its turn, it will enter the Battle state. If it is not its turn, it
will enter the Wait state.