feat: add multiplayer functionality
This commit adds the required functionality to enable more than one player to
play a battle of Ars Militaris.

Now, the server, upon receiving a connection from a client, will answer it with
that client's ClientId. This is necessary for the client to know its self ID.

The system assumes that the server has been is running and both clients are
connected (i.e., the three programs are at the MainMenu state), before issuing
the Start Game command.