feat: prepare for 0.1 release
This commit marks a state of the project where the basic architecture
and skeleton of Ars Militaris is implemented.

One or more `amclients` can connect either to the `amserver` running on
the Ars Militaris server, or connect to a local `amserver`. Right now
the change has to be done manually in the code but in the future a more
user-friendly switch will be provided.

Right now the client also contains a skeleton of a single-player scenario,
codenamed "Ambush". It is currently commented out in the code.

At the moment we're facing bug #6.