feat: load units into Ambush scenario
With this commit the functionality in the Ambush scenario has been
changed.

The cutscene_1 prototype which was being developed was deactivated,
and now the `spawn_units` system will read the
`the_patrol_ambush_data.csv` file, and spawn the units accordingly.

This commit also removes the `kafka` crate and its usage as it is
not necessary for now.