feat: add bevy_console and configure toggle key
This commit adds the crates `bevy_egui`, `bevy-inspector-egui`,
and `bevy_console` to the project.

It includes an initial configuration for `bevy_console`. With this
configuration, the player can now, at any time in the game, press
the "backslash" key, and the console will show up.

The objective with these 3 crates is to develop new editor
functionality, that will allows to develop game content more
productively.