feat: nothing console command
This commit adds the first command to the Ars Militaris console.

The `nothing` command is responsible for adding a DoNothing
UnitAction to the units. At the moment, it adds a DoNothing
UnitAction to every unit in the battle.

The command was developed followed the example at `bevy_console`
repository and crates.io page.