feat: complete processing of `Talk` `UnitAction`
This commit implements the missing functionality for correctly
processing a `Talk` `UnitAction`.

A new variant of the `GameState` enum, `GameState::Talk` was
created to support the new behavior.

Now, upon processing a `Talk` `UnitAction`, the game will wait for
the player to press the Left mouse button, to signal that the
player has read the message and the game can continue.

Also, the icons for the executable, window and taskbar were
updated to the new transparent logo.