fix: fix BUG#8
This commit fixes BUG#8, namely, the fact that the Linux build of
the game would crash upon trying to load the files in the `src`
directory.

This was due to the use of the `\\` as path separator, which is
Windows-only.

The path separator was changed to `/`, and now the game supports
all platforms.