feat: add demo prototype
Work has started on the Ars Militaris demo, which is to start
playing after some time when the player is idle on the Main Menu.

The demo is contained in the `ars_militaris_demo` system. For now,
it has two options:

1) Selects a random unit and inserts a random movement.
2) Selects the first unit and inserts a movement in the X
direction.

While developing this, work was also done on the first animation
of Ars Militaris: the animation for the `Move` `UnitAction`. Now,
upon movement, the unit will "slide" from its starting destination
to its target destination.

There are still some issues with the
z-ordering, and also the functionality of moving one tile at each
time isn't implemented yet.