feat: improve demo
With this commit we add the basic functionality for the demo to
select a unit and add a `Move` `UnitAction` to it.

Along the way we fixed the z-ordering of the units while moving.

We also created the `SinglePlayerPause` State, which is accessed
by pressing the 'F3' key (as inspired by Age of Empires), and
which pauses the game. This is very useful already for debugging.