feat: mechanism for team-2-player definition
This commit adds a mechanism for defining who is going to be the
controller of a certain team in an Ars Militaris battle.

This mechanism lives in the `Game` Resource, and will be able to
be interacted with with the WorldInspector and Ars Militaris
Console once the work to integrate with them has been done.

There are currently two options, defined by the `ControlledBy`
enum. `ControlledBy::Player` and `ControlledBy::AI`.

The game is now waiting for its first simple AI to be implemented.