feat: first AI
This commit adds the first AI to the game.

Right now, this AI is really simple and will only do a `Move`
`UnitAction` of a single tile in the `West` `Direction`, and then
end its turn.

But this means that the current implementation for the turns
mechanic is working and that it supports one (or more) AI players.