feat: current State visible with ResourceInspector
This commit adds two `ResourceInspector`s, one for each State
currently in the game.

For this, it was necessary to patch Bevy, more specifically the
`bevy_ecs` crate, in order to make the `State` type reflectable.

It is not recommended to change the current state using these
inspectors, as it will not trigger the `apply_state_transitions`
system of Bevy, and the `OnEnter()` and `OnExit()` schedules won't
be run.

Instead, these inspectors function more as a visibility tool to
check the current state of the game.