feat: attempt at inspecting NextState
We attempted to create a `ResourceInspector` for the `NextState`
Bevy Resource.

This `ResourceInspector` would allow us to modify the current
State, but unfortunately as it contains an `Option<State>`, we
haven't been able to deal with that yet, so currently it is not
possible to use this new `ResourceInspector`.

We leave the code here because it may be useful if we try to
develop a custom Inspector for the States.