feat: add new battle cursor
This commit adds the functionality for a new Cursor that the
player can control and which will be a central point of the
player's interaction with the game.

A new asset was generated for this Cursor, living in the
`cursor.png` file.

The Cursor follows the following implementation:

The Cursor is initialized in the `init_cursor_system` with its
Visibility set to Hidden.

When a player's turn begins, the `setup_cursor_system_2` runs,
and the Cursor's Visibility is set to Visible.

When the player's turn ends, the `hide_cursor` system runs and
sets the Visibility back to Hidden.

The player can control the Cursor using the 'WASD' keys.