feat: finish Move UnitAction handling
With this commit we complete the handling of the `Move`
`UnitAction`.

Now, when a `Move` `UnitAction` is added to a unit's `UnitAction`s
queue, it will then get divided into the individual `MoveAction`s
that are needed to make the desired movement (with a path computed
with an A* pathfinding algorithm provided by ChatGPT and the
`pathfinding` crate), and then the `handle_move_state` system
knows how to process each individual `MoveAction` in order.

There are some details that are not yet right with the
implementation, but this is already a great step forward.