feat: add Unit Directions
This commit implements the `Directions` feature of Ars Militaris.

Because Ars Militaris uses square tiles, there are four directions:
`East`, `South`, `West` and `North`. They are stored in the `enum`
`Direction`.

The `handle_unit_directions` system was created to update the unit
sprites when the direction changes.

Then in the `handle_move_state` system, changes were made so that,
when a unit is moving, it's direction is updated accordingly.