feat: add MoveTiles and moving with cursor
This commit adds a great deal of functionality.

In the first place, it adds the `ChooseMove` `TurnState`, to
handle the moment when a player is choosing the tile towards
which their unit will move.

The player can enter this State by pressing the 'M' key when upon
their turn. Upon doing this, the `MoveTiles` will show up, to
indicate the tiles that the unit can move to.

These tiles are calculated with the `find_possible_movements`
function.

The player can move the cursor to any of these tiles, and upon
pressing 'M' again, the unit will make a movement towards that
tile.

The game will then return to the `Turn` `TurnState` for the player
to issue more actions.

There is some polishing to be done on these new features, but
overall they are a solid development.