feat: show attack tiles
With this commit we begin to implement the combat system for
Ars Militaris.

The first thing in order was to have the available tiles for an
unit to attack, showing up with an indicator. That's what we have
done, using mostly the same architecture as we did for the `Move`
ability.

We develop an implementation for finding the available attack
tiles for both the Melee and Ranged AttackTypes.