feat: basic combat system
With this commit we complete a basic implementation of the combat
system for Ars Militaris.

Currently, it is possible for the player to issue a basic attack
command, using the Ars Militaris cursor and the 'F' key, and then
selecting a target.

Also implemented are the counterattacks feature of Ars Militaris.
Melee units will issue an automatic counterattack if at range.

The current battle system is very simple, and the basic attack
damage is calculated in the following way: it is the unit's STR
attribute divided by 3, then added or subtracted by a random value
between 0 and 3 (i.e., from -3 to 3, inclusive).

It was also implemented the feature so that units change their
direction when attacking an opponent.

Lastly, we note here that the current battle system is missing the
"block" feature, but that will come later.