fix: build_macos.yml syntax errors
This commit fixes the YAML syntax errors in the `build_macos`
The errors were being caused by the usage of tabs as indentation,
which is not allowed by YAML (it needs spaces as indentation).

We discovered this by using the tool YAML Lint
(https://www.yamllint.com/).

We hope that now the GitHub Action will run.