fix: fix macOS release
This commit adds a conditional configuration for the running of the
`set_window_icon` system. Now, this system will only run when the
host operating system is Windows.

This was necessary because the current Ars Militaris build for the
macOS platform was encountering an error on startup due to this
system, which should not be run under macOS.

As a note, by adding this conditional configuration, possibly the
Linux build will not display the window icon as well. We will test
this, and if so, find a solution that works for the current three
target platforms (Windows, macOS, Linux).