fix: fix build_macos.yml
The `build_macos.yml` GitHub Action was missing the step to copy the `src` directory into the app Bundle. 

This commit adds that step to the GitHub Action.

We don't know if when running on macOS the `amclient` will be able to detect the `src` directory. If it doesn't, we will move the assets that are there to the `assets` directory (which is where they deserve to be anyway).