fix: fix Multiplayer button
This commit fixes the issue where the "Multiplayer" scenario
wasn't loading after clicking the "Multiplayer" button in the
Main Menu.

The issue was that the code for that button didn't use the
necessary logic to connect to the server and send the
`ClientMessage::StartGame` message.

Now, the "Multiplayer" button correctly sends the message to the
server, and loads up the "Multiplayer" scenario correctly.