feat: update multiplayer
This commit updates the "Multiplayer" scenario, so that it loads
the same data as the "Ambush" single-player scenario.

Most of the systems from single-player were able to be used for
the multiplayer aspect, which is a nice thing about our
implementation.

Currently, the game is able to support two human players. But the
only Battle mechanic that is currently implemented is the ability
to end an unit's turn.

The next steps are to implement the networking of the `Move` and
`BasicAttack` `UnitActions`.