fix: fix multiplayer end turn
This commit fixes an issue with the end turn mechanic in the
multiplayer aspect of Ars Militaris.

The issue was that the game was using the `end_turn_single_player`
system when in the Multiplayer scenario, which was not correct.

Now, the Multiplayer scenario correctly uses the `end_turn_system`.

There's a slight bug, in that, when the player has finished all
their unit actions, and the game moves on to the next player, the
cursor won't be removed from the first player. Perhaps this is
desirable, so that the player can continue inspecting the map with
the cursor even when in another player's turn.