feat: add multiplayer Move UnitAction synchronization
This commit updates the multiplayer aspect of Ars Militaris with
the ability to correctly process `Move` `UnitAction`s sent by the
players.

Now, when a player issues a `Move` `UnitAction`, it will be sent
to the server, and the server will broadcast that `UnitAction` to
all the other clients. And, when it is processed, both (or more)
players will see the unit moving at the same time.

Now, the only thing missing to have the Multiplayer aspect of Ars
Militaris in a good state is to have the `BasicAttack` `UnitAction`
processed correctly by the server.