feat: add `BasicAttack` `UnitAction` synchronization
This commit adds the ability for, in an Ars Militaris multiplayer
battle, the `BasicAttack` `UnitAction` be synchronized between
server and clients.

Now, when a player issues a `BasicAttack` `UnitAction`, it will
get sent to the server, and the server will broadcast the
`BasicAttack` to all the clients, with the damage that was
calculate.

One learning was that the damage can only be calculated once, and
it should be in the server.

Now, the next step is to add the ability for the multiplayer
battle to detect when the game is over (currently, when every unit
in a team dies).