feat: add Game Over synchronization
This commits adds the ability for the server to detect when a
battle is over (i.e., when all the units in a team die), and end
the battle accordingly, sending the clients back to the Main Menu
with a message containing the winner.

This implementation is not completely tested but we think that it
will work.